src/monitor/requirements.txt
requests==2.22.0
python-dotenv==0.10.2
mysql-connector-python==8.0.16